vm: improvements
Pages are no longer added to the pages_t.head, instead pages_t kepps the
direct pointer page_t. Additionally:
* page.c: add pages_split function
* vm.c: vm_create no longer takes size argument; add vm_unmap_range
  function
* mmap.c: sbrk can now deallocate pages with negative incr
* select.c: fix leak in pselect/poll when when timeout is reached

+++ include/kernel/api/unistd.h
@@ -34,7 +34,7 @@ int reboot(int magic, int magic2, int cmd);
 int chdir(const char* buf);
 char* getcwd(char* buf, size_t size);
 int brk(void* addr);
-void* sbrk(size_t incr);
+void* sbrk(int incr);
 int chroot(const char* path);
 int setsid(void);
 int setuid(gid_t gid);
